instant-agents/tachi/weather-bot/requirements.txt
smolagents
python-dotenv
ddgs
